requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data Handling
pandas>=2.0.0

# Dashboard
streamlit>=1.28.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Progress Bars
tqdm>=4.65.0

# Configuration Management
pyyaml>=6.0

# EnergyPlus Integration
eppy>=0.5.63
# Note: EnergyPlus itself must be installed separately from https://energyplus.net/

# Jupyter for Research
jupyter>=1.0.0
ipykernel>=6.22.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
